==> results_scripts/2-datathon_tree_figure.Rmd <==
---
title: "Phylogenetic Tree of Datathon Taxa"
output: html_notebook
---

written by E.D. Crandall (eric.d.crandall@gmail.com, ORCID: 0000-0001-8580-3651)


I'm going to make a phylogenetic tree of the taxa that we addressed in the first datathon. Following along [here](https://mctavishlab.github.io/R_OpenTree_tutorials/01-find-taxa/index.html).


# Setup

```{r}
library(tidyverse)
library(rotl)
library(ape)


```


# The taxa

Right off the bat we find the problem that `specificEpithet` usually contains the generic name, but not always. I'm going to solve this manually, rather than programatically.

```{r}

biosamps <- read_tsv("datathon1_working_files/paper2_all_QCd_biosample_list.tsv",
                     col_types = cols(.default =   col_character()))

#get distinct genus/species
uniqueTaxa <- biosamps  %>% distinct(genus,specificEpithet)
# sub in genus where specificEpithet is missing
uniqueTaxa$specificEpithet[which(is.na(uniqueTaxa$specificEpithet))] <- uniqueTaxa$genus[which(is.na(uniqueTaxa$specificEpithet))]

#write_csv(uniqueTaxa, "datathon1_working_files/distinct_taxa_datathon1.csv")

#uniqueTaxa <- read.csv("datathon1_working_files/distinct_taxa_datathon1.csv")


```

Match them to the Open Tree of Life using the `rotl` package

```{r}


resolved_names <- tnrs_match_names(uniqueTaxa$specificEpithet, context_name = "All life", 
                                   include_suppressed = F)
# remove the NAs for 5 unmatched taxa: 
# Warning: Bufotes cypriensis, Bufotes perrini, Dactylorhiza pytagorae, Microbotryum lychnidis, Micromeria pedro are not matched
ott <- resolved_names %>% drop_na() %>% filter(!str_detect(flags,"incertae_sedis")) %>% 
                                          filter(!str_detect(flags,"extinct")) %>% 
                                          filter(!str_detect(flags,"hidden_inherited")) %>% 
                                          filter(!str_detect(flags,"hybrid")) %>%
                                          filter(!str_detect(flags,"pruned_ott_id")) %>%
                                          filter(!str_detect(flags,"major_rank_conflict")) %>%
                                          select(ott_id)
                                          

#remove `` additional taxa with IDs that cannot be found in the OTL tree
# Error: HTTP failure: 400
# [/v3/tree_of_life/induced_subtree] Error: node_id 'ott600790' was not found!list(ott600790 = "pruned_ott_id", ott877901 = "pruned_ott_id")
# these are extinct,  "hidden inherited", incertae_sedis, 


```

# Get a Subtree

```{r}
datathon_subtree <- tol_induced_subtree(ott_ids = ott$ott_id, label_format = "name", 
                                        file = "datathon_subtree.tre")
datathon_subtree <- tol_induced_subtree(ott_ids = ott$ott_id, label_format = "name")

write.tree(datathon_subtree,"datathon_subtree_nosingletons.tre")

# Warning: Dropping singleton nodes with labels: mrcaott2ott3973, mrcaott2ott276, mrcaott2ott142555, mrcaott2ott1551, mrcaott2ott7623, Streptophyta, mrcaott2ott50189, mrcaott2ott108668, mrcaott2ott59852, mrcaott2ott8171, Embryophyta, mrcaott2ott70394, Magnoliopsida, mrcaott2ott2645, mrcaott2ott35778, mrcaott2ott10930, mrcaott2ott969, mrcaott2ott62529, mrcaott2ott8379, eudicotyledons, Gunneridae, mrcaott2ott8384, fabids, mrcaott2ott1479, mrcaott2ott345, Malpighiales, mrcaott345ott3853, mrcaott345ott3949, mrcaott345ott22364, mrcaott345ott18689, mrcaott345ott177705, mrcaott345ott38722, mrcaott345ott24237, mrcaott345ott30307, mrcaott345ott26478, mrcaott345ott191443, mrcaott345ott6748, mrcaott345ott504, mrcaott345ott485790, mrcaott345ott30323, Euphorbieae, mrcaott345ott1030065, Euphorbia, mrcaott14597ott33281, mrcaott14597ott22595, mrcaott14597ott14599, mrcaott14597ott70724, mrcaott371ott773, mrcaott371ott63303, mrcaott371ott22991, Moraceae, mrcaott371ott34788, mrcaott371ott35317, mrcaott371ott145210, Cannabaceae, mrcaott74296ott833737, mrcaott74296ott113320, mrcaott74296ott494614, mrcaott74296ott348783, mrcaott74296ott74299, mrcaott74299ott84002, mrcaott74299ott582836, mrcaott74299ott4729332, Cannabis, Rosoideae, mrcaott1448ott18466, mrcaott1448ott28620, mrcaott1448ott32041, mrcaott32041ott32065, Potentilleae, Fragariinae, mrcaott355554ott1031262, mrcaott355554ott427871, mrcaott355554ott1004788, mrcaott10421ott122440, Amygdaloideae, mrcaott10421ott34671, mrcaott10421ott156934, mrcaott10421ott22223, mrcaott10421ott32528, mrcaott10421ott873692, Maleae, mrcaott10421ott366133, mrcaott10421ott1085086, mrcaott10421ott22639, mrcaott22639ott62443, mrcaott22639ott846359, mrcaott22639ott29980, mrcaott29980ott739733, mrcaott29980ott52112, mrcaott29980ott153973, mrcaott29980ott382000, mrcaott45825ott52104, mrcaott60966ott582042, mrcaott60966ott60983, mrcaott60983ott547120, mrcaott60983ott653070, mrcaott60983ott129779, mrcaott2511ott32687, Fagales, mrcaott32687ott47808, mrcaott47808ott70096, Juglandaceae, mrcaott70096ott199041, mrcaott70096ott199032, mrcaott70096ott264610, mrcaott70096ott1091658, mrcaott70096ott101820, mrcaott70096ott70104, mrcaott70096ott1072887, mrcaott37377ott123901, mrcaott37377ott81721, mrcaott37377ott71501, mrcaott37377ott137331, mrcaott37377ott257677, mrcaott37377ott528690, mrcaott37377ott1062902, mrcaott37377ott117129, mrcaott37377ott106847, mrcaott37377ott381049, mrcaott37377ott288806, mrcaott37377ott3930928, mrcaott37377ott288802, mrcaott37377ott61992, mrcaott37377ott316112, mrcaott37377ott106849, mrcaott37377ott554707, mrcaott554707ott1056277, mrcaott1056277ott7056093, mrcaott252507ott403395, mrcaott117134ott288819, mrcaott117134ott272703, mrcaott1026065ott3930866, Fagus, mrcaott288140ott991650, mrcaott288140ott685939, mrcaott685939ott991641, Fabales, mrcaott579ott179386, Fabaceae, mrcaott579ott45621, mrcaott579ott8603, Papilionoideae, mrcaott579ott5464, mrcaott579ott5457, mrcaott579ott5462, mrcaott579ott3942, mrcaott579ott5470, mrcaott579ott5091, mrcaott579ott34386, mrcaott579ott38037, mrcaott579ott7223, mrcaott579ott11248, mrcaott11248ott128412, mrcaott11248ott11254, mrcaott1749ott496549, mrcaott1749ott222250, mrcaott1749ott27613, mrcaott1749ott38031, mrcaott38031ott42707, mrcaott38031ott38042, Medicago, mrcaott38031ott209352, mrcaott38031ott424990, mrcaott38031ott38032, mrcaott38031ott525750, mrcaott38031ott193971, mrcaott38031ott760436, mrcaott38031ott1007991, mrcaott5515ott44910, mrcaott5515ott170171, mrcaott5515ott48574, mrcaott5515ott16972, mrcaott16972ott117879, mrcaott16972ott931944, mrcaott16972ott997717, mrcaott16972ott222824, mrcaott16972ott33188, mrcaott16972ott131682, mrcaott16972ott569265, mrcaott16972ott73347, mrcaott16972ott80721, mrcaott16972ott21757, mrcaott21757ott34156, mrcaott21757ott116125, mrcaott21757ott116123, mrcaott21757ott219566, mrcaott21757ott48568, mrcaott21757ott161299, Glycine, mrcaott108936ott459708, Soja, mrcaott119902ott213839, Amphicarpaea, mrcaott2644ott112865, mrcaott2644ott86941, mrcaott2644ott8605, mrcaott8605ott87214, mrcaott8605ott21076, mrcaott8605ott220379, mrcaott8605ott40823, mrcaott8605ott308458, mrcaott8605ott659681, mrcaott8605ott26386, mrcaott8605ott8606, mrcaott8878ott17129, mrcaott8878ott444652, mrcaott8878ott146599, mrcaott8878ott33836, mrcaott8878ott8883, mrcaott8878ott8880, mrcaott8878ott9304, mrcaott9304ott337248, mrcaott9304ott198224, mrcaott9304ott80192, mrcaott9304ott996250, mrcaott9304ott9318, mrcaott9318ott97447, mrcaott9318ott3917743, mrcaott9318ott274403, mrcaott9318ott3918125, mrcaott9318ott97454, mrcaott9318ott46049, mrcaott9318ott53279, mrcaott9318ott554919, mrcaott9318ott39866, mrcaott39866ott3918261, mrcaott39866ott188178, mrcaott39866ott4730372, mrcaott60608ott188184, mrcaott60608ott320032, mrcaott320032ott3917761, mrcaott320032ott867933, mrcaott320032ott535353, mrcaott320032ott633791, mrcaott633791ott3917644, mrcaott3917644ott3918447, mrcaott730ott15950, Detarioideae, mrcaott730ott43631, mrcaott730ott31285, mrcaott730ott82405, mrcaott730ott228631, Afzelieae, mrcaott228631ott301494, Afzelia (genus in kingdom Archaeplastida), mrcaott96ott14140, mrcaott96ott50744, Sapindales, mrcaott96ott84975, mrcaott96ott655994, mrcaott96ott21231, Burseraceae, mrcaott96ott132934, mrcaott132934ott257009, mrcaott132934ott368075, mrcaott1860ott9337, mrcaott9337ott956950, Sapindaceae, mrcaott9337ott23895, mrcaott9337ott81662, mrcaott81662ott184020, mrcaott81662ott143045, mrcaott81662ott237222, mrcaott81662ott468155, mrcaott81662ott217381, mrcaott81662ott162812, mrcaott81662ott981704, mrcaott81662ott88893, mrcaott81662ott315099, mrcaott81662ott88887, mrcaott88887ott811780, mrcaott88887ott248950, mrcaott88887ott811778, mrcaott88887ott162803, mrcaott88887ott318105, mrcaott88887ott88891, mrcaott88887ott248952, mrcaott88887ott374653, mrcaott88887ott315109, mrcaott88887ott88889, mrcaott88887ott957601, mrcaott378ott29446, Brassicales, mrcaott378ott307071, mrcaott378ott32461, mrcaott378ott509555, mrcaott378ott318175, mrcaott378ott9635, mrcaott378ott125843, mrcaott378ott509568, mrcaott378ott28763, mrcaott378ott83547, mrcaott378ott299734, mrcaott378ott4671, mrcaott4671ott58909, mrcaott4671ott6278, mrcaott6278ott15318, mrcaott6278ott158438, mrcaott6278ott10585, mrcaott6278ott34460, mrcaott6278ott193211, Erysimeae, Erysimum, mrcaott9083ott19798, mrcaott19798ott682496, mrcaott42848ott991635, mrcaott31487ott88883, mrcaott31487ott152275, mrcaott152275ott220174, mrcaott265527ott5785236, mrcaott265527ott433348, mrcaott11023ott24850, mrcaott11023ott56298, mrcaott11023ott95692, mrcaott95692ott440662, mrcaott95692ott185264, Malvales, mrcaott1697ott3190, mrcaott3190ott103459, Malvaceae, mrcaott3190ott21755, mrcaott3190ott108617, mrcaott3190ott17908, mrcaott3190ott388185, mrcaott3190ott336233, mrcaott3190ott103330, mrcaott3190ott123043, mrcaott3190ott5545, mrcaott3190ott4358, mrcaott3190ott5548, mrcaott3190ott253263, mrcaott3190ott286157, mrcaott3190ott236636, mrcaott3190ott4363, mrcaott23117ott538376, mrcaott23117ott84580, mrcaott84580ott642479, mrcaott84580ott779801, mrcaott779801ott779803, mrcaott779801ott779808, mrcaott8010ott287352, mrcaott8010ott19033, mrcaott19033ott762061, Dipterocarpaceae, mrcaott19033ott463925, mrcaott19033ott691511, mrcaott19033ott291594, mrcaott19033ott990342, mrcaott19033ott19048, mrcaott19033ott1066661, mrcaott19033ott19040, mrcaott19033ott19036, mrcaott19036ott116984, mrcaott19036ott30486, mrcaott19036ott343129, mrcaott19036ott861298, mrcaott19036ott30488, mrcaott19036ott19038, mrcaott19038ott125801, mrcaott125801ott540289, mrcaott125801ott603751, mrcaott125801ott603745, mrcaott607ott1276, Myrtales, mrcaott1276ott2663, mrcaott2663ott45629, Myrtaceae, mrcaott2663ott133747, mrcaott2663ott2665, mrcaott2665ott29679, mrcaott2665ott9519, mrcaott9519ott9527, mrcaott9527ott484459, mrcaott9527ott13016, Eucalypteae, mrcaott13016ott39207, Eucalyptus, mrcaott13016ott173803, mrcaott13016ott159098, mrcaott13016ott847939, mrcaott13016ott179898, mrcaott179898ott191238, mrcaott179898ott6121998, mrcaott179898ott275275, mrcaott179898ott278186, mrcaott278186ott325236, mrcaott278186ott748054, mrcaott278186ott473812, mrcaott278186ott473807, mrcaott278186ott325231, Saxifragales, mrcaott2464ott14587
```

Ended up using this tree for the figure:
Phylogenetic tree with 672 tips and 581 internal nodes.

Tip labels:
  Euphorbia_niciciana, Euphorbia_nicaeensis, Euphorbia_seguieriana, Ficus_arfakensis, Ficus_microdictya, Ficus_itoana, ...
Node labels:
  Eukaryota, mrcaott2ott148, Chloroplastida, Tracheophyta, Euphyllophyta, Spermatophyta, ...

Rooted; no branch lengths.
